refactor(core): split favicon pipeline from custom icon storage
The favicon pipeline and custom-icon storage shared one 1304-line file but
solved unrelated problems: one is a network-sourced acquisition pipeline
(URL → hosts → candidates → HTTP fetch → validate → normalize), the other
is KDBX metadata CRUD on `db.meta.custom_icons` with content-hash dedup.

Splits into two modules with separate test surfaces:

- `services/kdbx/favicons/` — `lookup`, `candidates`, `http`, `image`
  submodules. Public seam is `fetch_favicon_for_url(...)` returning a
  `FaviconFetchResult` enum. The pipeline no longer touches the Vault;
  the caller decides how to reconcile bytes with the Entry's icon state.
- `services/kdbx/custom_icons.rs` — owns `set_/clear_/assign_entry_custom_icon`,
  `get_custom_icons`, `detect_icon_mime`, and the hash-dedup logic.

`FaviconCooldown` becomes its own type held by `KdbxService` instead of
inline `Mutex<HashMap<...>>` fields and free methods. Cooldown tests
talk to it through `seed_for_test`/`stamp_for_test` rather than reaching
into `service.favicon_failed_domains.lock()`.

`KdbxService::fetch_entry_favicon` is now a ~30-line coordinator: locked
read of the URL, drop the lock, run the unlocked pipeline, then take a
second lock to write bytes (via `assign_entry_custom_icon`) and update
cooldown. The lock dance is now explicit instead of smeared across one
118-line function.

IPC surface (`fetch_entry_favicon`, `set_entry_custom_icon`,
`clear_entry_custom_icon`, `get_custom_icons`, `FaviconFetchOutcome`)
is unchanged — frontend is unaffected.

Also seeds `CONTEXT.md` with the load-bearing domain terms (Vault, Entry,
Group, Custom Icon, Favicon, Favicon Lookup, Favicon Candidate, Favicon
Cooldown, Favicon Fetch Result, Favicon Fetch Outcome, App Preferences,
Database Config, Secure String/Bytes).
