+++ CONTEXT.md
@@ -1,0 +1,54 @@
+# MithrilVault — Domain Language
+
+This file names the concepts the codebase is *about*. Terms here are load-bearing — they appear in module names, function names, and conversations. If a term drifts, fix it here first.
+
+Architecture vocabulary (Module, Interface, Seam, Adapter, Depth) is separate and lives in `.claude/skills/improve-codebase-architecture/LANGUAGE.md`.
+
+## Storage
+
+### Vault
+A KDBX-format password database (KDBX3 or KDBX4) that MithrilVault has opened. Identified by its filesystem path. May be **locked** (encrypted, no decrypted state in memory) or **unlocked** (decrypted KDBX tree is live in memory and editable). Multiple Vaults can be open at once, keyed by path.
+
+### Entry
+A single credential record inside a Vault. Has a title, URL, username, password, optional notes, tags, custom fields, and an icon. Password material is read separately from the rest (see CLAUDE.md "minimal data in list views").
+
+### Group
+A folder inside a Vault that contains Entries and other Groups. Forms a tree rooted at the Vault's root Group.
+
+## Icons
+
+KDBX 0.12+ makes built-in and custom icons mutually exclusive on an Entry — an Entry has either an Icon ID or a Custom Icon UUID, never both.
+
+### Custom Icon
+A binary image stored inside the Vault's metadata (`db.meta.custom_icons`), referenced from an Entry by UUID. Deduped within a Vault by content hash (Sha256). MIME-aware over IPC (`{ mimeType, data }`).
+
+### Favicon
+A web-sourced icon for an Entry, derived from the Entry's URL. When a favicon is downloaded, it is **stored as a Custom Icon** — there is no separate "favicon" storage. "Favicon" describes the *acquisition pipeline*, not the storage.
+
+### Favicon Lookup
+The host derivations extracted from an Entry's URL, used to enumerate where to fetch a Favicon from. Currently three fields: the exact authority (hostname + port), the hostname alone (lowercased), and the registrable root domain (PSL-derived eTLD+1, if any).
+
+### Favicon Candidate
+One attempt in the Favicon pipeline: a `fetch_url` to try and a `cooldown_domain` to record success/failure against. Multiple Candidates are derived from one Favicon Lookup — direct host, root host, and (opt-in) third-party services like Google or icon.horse.
+
+### Favicon Cooldown
+A per-process, per-`cooldown_domain` failure timestamp. Used to skip Candidates that recently failed, so a bulk auto-fetch doesn't hammer the same dead host across every Entry that uses it. Bypassed by the manual "Refetch" path (`force=true`).
+
+### Favicon Fetch Result
+The public output of the Favicon pipeline. Either `Found { bytes, mime_type, cooldown_domain }` (a successful Candidate produced usable image bytes) or `NotFound { attempted_domains }` (every Candidate failed or was skipped). The pipeline does **not** write to the Vault — the caller decides what to do with the bytes.
+
+### Favicon Fetch Outcome
+The Entry-level outcome of `fetch_entry_favicon`: `Updated`, `Unchanged`, or `NotFound`. Distinct from Favicon Fetch Result — this is what the IPC layer surfaces to the UI after the pipeline result has been reconciled with the Entry's existing icon state.
+
+## Settings
+
+### App Preferences
+The editable user-facing settings (theme, language, clipboard timeout, screen-capture protection, favicon auto-fetch policy). Stored locally per machine. Hierarchical shape on the IPC boundary; persisted as a flat `AppSettings` shape.
+
+### Database Config
+A read-only snapshot of properties intrinsic to a specific Vault (KDF parameters, KDBX version, generator string). Surfaced by `get_database_config`. Not editable through the same path as App Preferences.
+
+## Security
+
+### Secure String / Secure Bytes
+Wrappers around `String` / `Vec<u8>` that zeroize on drop and redact in Debug/Display output. Used for every value the code wants to keep out of accidental logs and core dumps — passwords, keyfile contents, derived keys.
