--- src-tauri/src/services/kdbx/favicons.rs
@@ -1,1304 +1,0 @@
-use crate::dto::error::AppError;
-use image::imageops::FilterType;
-use image::ImageFormat;
-use keepass::db::{Icon, Times};
-use reqwest::header::CONTENT_TYPE;
-use reqwest::redirect::Policy;
-use serde::Serialize;
-use sha2::{Digest, Sha256};
-use std::collections::HashSet;
-use std::io::Cursor;
-use url::Url;
-use uuid::Uuid;
-
-use super::mapping::find_entry_id;
-use super::KdbxService;
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize)]
-#[serde(rename_all = "camelCase")]
-pub enum FaviconFetchOutcome {
-    Updated,
-    Unchanged,
-    NotFound,
-}
-
-const FAVICON_MAX_BYTES: usize = 512 * 1024;
-const GOOGLE_FAVICON_URL: &str = "https://www.google.com/s2/favicons";
-const ICON_HORSE_URL: &str = "https://icon.horse/icon";
-
-impl KdbxService {
-    pub async fn fetch_entry_favicon(
-        &self,
-        db_id: &str,
-        entry_id: &str,
-        allow_third_party_fallbacks: bool,
-        force: bool,
-    ) -> Result<FaviconFetchOutcome, AppError> {
-        let (entry_url, has_custom_icon) = {
-            let normalized_path = Self::normalize_path(db_id);
-            let databases = self.lock_databases()?;
-            let open_db = databases
-                .get(&normalized_path)
-                .ok_or_else(|| AppError::DatabaseNotFound(db_id.to_string()))?;
-            let db = open_db.db_or_locked()?;
-            let eid = find_entry_id(db, entry_id)
-                .ok_or_else(|| AppError::EntryNotFound(entry_id.to_string()))?;
-            let entry = db
-                .entry(eid)
-                .ok_or_else(|| AppError::EntryNotFound(entry_id.to_string()))?;
-            (
-                entry.get_url().map(str::to_string),
-                matches!(entry.icon(), Some(Icon::Custom(_))),
-            )
-        };
-
-        if !force && has_custom_icon {
-            return Ok(FaviconFetchOutcome::Unchanged);
-        }
-
-        let Some(raw_url) = entry_url else {
-            return Ok(FaviconFetchOutcome::NotFound);
-        };
-
-        let Some(hosts) = extract_hosts(&raw_url) else {
-            return Ok(FaviconFetchOutcome::NotFound);
-        };
-
-        let client = reqwest::Client::builder()
-            .connect_timeout(std::time::Duration::from_secs(5))
-            .timeout(std::time::Duration::from_secs(10))
-            .redirect(Policy::custom(|attempt| {
-                if attempt.url().scheme() != "https" {
-                    attempt.stop()
-                } else if attempt.previous().len() >= 5 {
-                    attempt.error("too many redirects")
-                } else {
-                    attempt.follow()
-                }
-            }))
-            .user_agent("MithrilVault/0.1")
-            .build()
-            .map_err(|error| AppError::Io(error.to_string()))?;
-
-        let candidates = build_favicon_candidates(&hosts, allow_third_party_fallbacks);
-        let mut attempted_domains = HashSet::new();
-
-        for candidate in candidates {
-            if !force && self.is_favicon_domain_on_cooldown(&candidate.cooldown_domain)? {
-                continue;
-            }
-            attempted_domains.insert(candidate.cooldown_domain.clone());
-
-            let Some((downloaded_bytes, content_type)) =
-                fetch_favicon_bytes(&client, &candidate.fetch_url).await
-            else {
-                continue;
-            };
-
-            let (icon_bytes, mime_type) = normalize_favicon_bytes(&downloaded_bytes, content_type);
-            if icon_bytes.is_empty() {
-                continue;
-            }
-
-            let changed =
-                self.assign_entry_custom_icon(db_id, entry_id, &icon_bytes, &mime_type, force)?;
-            let _ = self.clear_favicon_domain_failure(&candidate.cooldown_domain);
-            return Ok(if changed {
-                FaviconFetchOutcome::Updated
-            } else {
-                FaviconFetchOutcome::Unchanged
-            });
-        }
-
-        for domain in attempted_domains {
-            let _ = self.mark_favicon_domain_failed(&domain);
-        }
-
-        Ok(FaviconFetchOutcome::NotFound)
-    }
-
-    pub fn set_entry_custom_icon(
-        &self,
-        db_id: &str,
-        entry_id: &str,
-        icon_uuid: &str,
-    ) -> Result<bool, AppError> {
-        let parsed_uuid = Uuid::parse_str(icon_uuid)
-            .map_err(|error| AppError::InvalidInput(error.to_string()))?;
-
-        let normalized_path = Self::normalize_path(db_id);
-        let mut databases = self.lock_databases()?;
-        let open_db = databases
-            .get_mut(&normalized_path)
-            .ok_or_else(|| AppError::DatabaseNotFound(db_id.to_string()))?;
-        let db = open_db.db_mut_or_locked()?;
-
-        let icon_cid = db
-            .iter_all_custom_icons()
-            .find(|icon| icon.id().uuid() == parsed_uuid)
-            .map(|icon| icon.id());
-        let Some(icon_cid) = icon_cid else {
-            return Err(AppError::InvalidInput(format!(
-                "custom icon {icon_uuid} not found in database"
-            )));
-        };
-
-        let eid = find_entry_id(db, entry_id)
-            .ok_or_else(|| AppError::EntryNotFound(entry_id.to_string()))?;
-        let mut entry = db
-            .entry_mut(eid)
-            .ok_or_else(|| AppError::EntryNotFound(entry_id.to_string()))?;
-
-        if matches!(entry.icon(), Some(Icon::Custom(cid)) if *cid == icon_cid) {
-            return Ok(false);
-        }
-
-        entry
-            .set_icon_custom(icon_cid)
-            .map_err(|e| AppError::Kdbx(e.to_string()))?;
-        entry.times.last_modification = Some(Times::now());
-        open_db.is_modified = true;
-        Ok(true)
-    }
-
-    pub fn clear_entry_custom_icon(&self, db_id: &str, entry_id: &str) -> Result<bool, AppError> {
-        let normalized_path = Self::normalize_path(db_id);
-        let mut databases = self.lock_databases()?;
-        let open_db = databases
-            .get_mut(&normalized_path)
-            .ok_or_else(|| AppError::DatabaseNotFound(db_id.to_string()))?;
-        let db = open_db.db_mut_or_locked()?;
-
-        let eid = find_entry_id(db, entry_id)
-            .ok_or_else(|| AppError::EntryNotFound(entry_id.to_string()))?;
-        let mut entry = db
-            .entry_mut(eid)
-            .ok_or_else(|| AppError::EntryNotFound(entry_id.to_string()))?;
-
-        if !matches!(entry.icon(), Some(Icon::Custom(_))) {
-            return Ok(false);
-        }
-
-        entry.set_icon_none();
-        entry.times.last_modification = Some(Times::now());
-        open_db.is_modified = true;
-        Ok(true)
-    }
-
-    fn assign_entry_custom_icon(
-        &self,
-        db_id: &str,
-        entry_id: &str,
-        icon_bytes: &[u8],
-        _mime_type: &str,
-        force: bool,
-    ) -> Result<bool, AppError> {
-        let normalized_path = Self::normalize_path(db_id);
-        let mut databases = self.lock_databases()?;
-        let open_db = databases
-            .get_mut(&normalized_path)
-            .ok_or_else(|| AppError::DatabaseNotFound(db_id.to_string()))?;
-
-        let db = open_db.db_mut_or_locked()?;
-
-        let eid = find_entry_id(db, entry_id)
-            .ok_or_else(|| AppError::EntryNotFound(entry_id.to_string()))?;
-
-        let already_has = matches!(
-            db.entry(eid).and_then(|e| e.icon().cloned()),
-            Some(Icon::Custom(_))
-        );
-
-        if !force && already_has {
-            return Ok(false);
-        }
-
-        let target_hash = hash_bytes(icon_bytes);
-        let existing_cid = db
-            .iter_all_custom_icons()
-            .find(|icon| hash_bytes(&icon.data) == target_hash)
-            .map(|icon| icon.id());
-
-        let mut entry = db
-            .entry_mut(eid)
-            .ok_or_else(|| AppError::EntryNotFound(entry_id.to_string()))?;
-
-        match existing_cid {
-            Some(cid) => {
-                if matches!(entry.icon(), Some(Icon::Custom(current)) if *current == cid) {
-                    return Ok(false);
-                }
-                entry
-                    .set_icon_custom(cid)
-                    .map_err(|e| AppError::Kdbx(e.to_string()))?;
-            }
-            None => {
-                entry.set_icon_custom_new(icon_bytes.to_vec());
-            }
-        }
-
-        entry.times.last_modification = Some(Times::now());
-        open_db.is_modified = true;
-        Ok(true)
-    }
-}
-
-#[derive(Debug, Clone)]
-struct FaviconCandidate {
-    fetch_url: String,
-    cooldown_domain: String,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-struct FaviconLookup {
-    exact_authority: String,
-    hostname: String,
-    root_domain: Option<String>,
-}
-
-fn build_favicon_candidates(
-    hosts: &FaviconLookup,
-    allow_third_party_fallbacks: bool,
-) -> Vec<FaviconCandidate> {
-    let mut direct_hosts = vec![(hosts.exact_authority.as_str(), hosts.hostname.as_str())];
-    if let Some(root) = hosts.root_domain.as_deref() {
-        if root != hosts.hostname {
-            direct_hosts.push((root, root));
-        }
-    }
-
-    let mut candidates = Vec::new();
-    for (fetch_host, cooldown_domain) in &direct_hosts {
-        candidates.push(FaviconCandidate {
-            fetch_url: format!("https://{fetch_host}/favicon.ico"),
-            cooldown_domain: (*cooldown_domain).to_string(),
-        });
-    }
-
-    if allow_third_party_fallbacks {
-        let mut third_party_domains = Vec::new();
-        if let Some(domain) = get_public_registrable_domain(&hosts.hostname) {
-            third_party_domains.push(domain);
-        }
-        if let Some(root) = hosts.root_domain.as_deref() {
-            if let Some(domain) = get_public_registrable_domain(root) {
-                if !third_party_domains.contains(&domain) {
-                    third_party_domains.push(domain);
-                }
-            }
-        }
-
-        for domain in &third_party_domains {
-            candidates.push(FaviconCandidate {
-                fetch_url: format!("{GOOGLE_FAVICON_URL}?domain={domain}&sz=64"),
-                cooldown_domain: domain.clone(),
-            });
-            candidates.push(FaviconCandidate {
-                fetch_url: format!("{ICON_HORSE_URL}/{domain}"),
-                cooldown_domain: domain.clone(),
-            });
-        }
-    }
-
-    candidates
-}
-
-fn extract_hosts(entry_url: &str) -> Option<FaviconLookup> {
-    let parsed = Url::parse(entry_url).ok()?;
-    match parsed.scheme() {
-        "http" | "https" => {}
-        _ => return None,
-    }
-    let host = parsed.host_str()?.to_ascii_lowercase();
-    let exact_authority = format_fetch_host(&host, parsed.port());
-    let root = get_root_domain(&host);
-    Some(FaviconLookup {
-        exact_authority,
-        hostname: host,
-        root_domain: root,
-    })
-}
-
-fn format_fetch_host(host: &str, port: Option<u16>) -> String {
-    let host_for_url = if host.parse::<std::net::Ipv6Addr>().is_ok() {
-        format!("[{host}]")
-    } else {
-        host.to_string()
-    };
-
-    if let Some(port) = port {
-        format!("{host_for_url}:{port}")
-    } else {
-        host_for_url
-    }
-}
-
-fn get_root_domain(host: &str) -> Option<String> {
-    let registrable = get_public_registrable_domain(host)?;
-    if registrable.eq_ignore_ascii_case(host) {
-        return None;
-    }
-    Some(registrable)
-}
-
-fn get_public_registrable_domain(host: &str) -> Option<String> {
-    if host.parse::<std::net::Ipv4Addr>().is_ok() || host.parse::<std::net::Ipv6Addr>().is_ok() {
-        return None;
-    }
-
-    // Only allow hosts whose suffix is in the Public Suffix List. PSL marks
-    // suffixes derived via the implicit "*" wildcard rule (i.e. unknown
-    // TLDs like .local, .internal, or any private/special-use name) as
-    // unknown — those must not be sent to third-party favicon services.
-    if !psl::suffix(host.as_bytes())?.is_known() {
-        return None;
-    }
-
-    // For PSL-aware hosts, take the registrable domain (eTLD+1). When the
-    // host already IS the registrable apex (example.co.uk) or sits on a
-    // PSL platform suffix (user.github.io, app.netlify.app), domain_str
-    // returns the host itself and the caller's equality check skips the
-    // root-host fallback.
-    psl::domain_str(host).map(str::to_ascii_lowercase)
-}
-
-async fn fetch_favicon_bytes(
-    client: &reqwest::Client,
-    fetch_url: &str,
-) -> Option<(Vec<u8>, Option<String>)> {
-    let mut response = client.get(fetch_url).send().await.ok()?;
-    let requested_https = Url::parse(fetch_url).is_ok_and(|url| url.scheme() == "https");
-    if requested_https && response.url().scheme() != "https" {
-        return None;
-    }
-
-    if !response.status().is_success() {
-        return None;
-    }
-
-    let content_type = response
-        .headers()
-        .get(CONTENT_TYPE)
-        .and_then(|value| value.to_str().ok())
-        .map(normalize_content_type);
-
-    if let Some(ref value) = content_type {
-        if !is_potential_image_content_type(value) {
-            return None;
-        }
-    }
-
-    if let Some(len) = response.content_length() {
-        if len > FAVICON_MAX_BYTES as u64 {
-            return None;
-        }
-    }
-
-    let mut bytes: Vec<u8> = Vec::new();
-    loop {
-        match response.chunk().await {
-            Ok(Some(chunk)) => {
-                if bytes.len().saturating_add(chunk.len()) > FAVICON_MAX_BYTES {
-                    return None;
-                }
-                bytes.extend_from_slice(&chunk);
-            }
-            Ok(None) => break,
-            Err(_) => return None,
-        }
-    }
-
-    if bytes.is_empty() {
-        return None;
-    }
-
-    if !has_known_image_signature(&bytes) && !looks_like_svg(&bytes) {
-        return None;
-    }
-
-    Some((bytes, content_type))
-}
-
-fn normalize_content_type(value: &str) -> String {
-    value
-        .split(';')
-        .next()
-        .map(str::trim)
-        .map(str::to_ascii_lowercase)
-        .unwrap_or_default()
-}
-
-fn is_potential_image_content_type(content_type: &str) -> bool {
-    content_type.starts_with("image/")
-        || content_type == "application/xml"
-        || content_type == "text/xml"
-        || content_type == "application/octet-stream"
-}
-
-fn normalize_favicon_bytes(bytes: &[u8], content_type: Option<String>) -> (Vec<u8>, String) {
-    if let Ok(image) = image::load_from_memory(bytes) {
-        let resized = image.resize(64, 64, FilterType::Lanczos3);
-        let mut output = Cursor::new(Vec::new());
-        if resized.write_to(&mut output, ImageFormat::Png).is_ok() {
-            return (output.into_inner(), "image/png".to_string());
-        }
-    }
-
-    let mime = content_type.unwrap_or_else(|| detect_icon_mime(bytes));
-    (bytes.to_vec(), mime)
-}
-
-pub(crate) fn detect_icon_mime(bytes: &[u8]) -> String {
-    if looks_like_svg(bytes) {
-        return "image/svg+xml".to_string();
-    }
-
-    if bytes.starts_with(&[0x89, b'P', b'N', b'G']) {
-        return "image/png".to_string();
-    }
-    if bytes.len() >= 3 && bytes[0] == 0xFF && bytes[1] == 0xD8 && bytes[2] == 0xFF {
-        return "image/jpeg".to_string();
-    }
-    if bytes.starts_with(b"GIF87a") || bytes.starts_with(b"GIF89a") {
-        return "image/gif".to_string();
-    }
-    if bytes.starts_with(&[0x00, 0x00, 0x01, 0x00]) {
-        return "image/x-icon".to_string();
-    }
-    if bytes.starts_with(b"BM") {
-        return "image/bmp".to_string();
-    }
-    if bytes.starts_with(b"RIFF") && bytes.len() > 11 && &bytes[8..12] == b"WEBP" {
-        return "image/webp".to_string();
-    }
-    if bytes.starts_with(&[0x49, 0x49, 0x2A, 0x00]) || bytes.starts_with(&[0x4D, 0x4D, 0x00, 0x2A])
-    {
-        return "image/tiff".to_string();
-    }
-
-    "application/octet-stream".to_string()
-}
-
-fn has_known_image_signature(bytes: &[u8]) -> bool {
-    bytes.starts_with(&[0x89, b'P', b'N', b'G'])
-        || (bytes.len() >= 3 && bytes[0] == 0xFF && bytes[1] == 0xD8 && bytes[2] == 0xFF)
-        || bytes.starts_with(b"GIF87a")
-        || bytes.starts_with(b"GIF89a")
-        || bytes.starts_with(&[0x00, 0x00, 0x01, 0x00])
-        || bytes.starts_with(b"BM")
-        || (bytes.starts_with(b"RIFF") && bytes.len() > 11 && &bytes[8..12] == b"WEBP")
-        || bytes.starts_with(&[0x49, 0x49, 0x2A, 0x00])
-        || bytes.starts_with(&[0x4D, 0x4D, 0x00, 0x2A])
-}
-
-fn looks_like_svg(bytes: &[u8]) -> bool {
-    let Ok(text) = std::str::from_utf8(bytes) else {
-        return false;
-    };
-    text.trim_start().starts_with("<svg") || text.contains("<svg")
-}
-
-fn hash_bytes(bytes: &[u8]) -> [u8; 32] {
-    let mut hasher = Sha256::new();
-    hasher.update(bytes);
-    hasher.finalize().into()
-}
-
-#[cfg(test)]
-#[allow(clippy::expect_used)]
-mod tests {
-    use super::*;
-    use crate::domain::secure::SecureString;
-    use crate::dto::database::DatabaseCreationOptions;
-    use crate::dto::entry::CreateEntryData;
-    use image::{ImageBuffer, Rgba};
-    use std::io::{Read, Write};
-    use std::net::TcpListener;
-    use std::thread;
-    use tempfile::TempDir;
-
-    fn create_test_database() -> (KdbxService, TempDir, String, String, String) {
-        let dir = tempfile::tempdir().expect("create temp dir");
-        let db_path = dir.path().join("favicon-tests.kdbx");
-        let db_path_str = db_path.to_string_lossy().to_string();
-
-        let options = DatabaseCreationOptions {
-            create_default_groups: true,
-            kdf_memory: Some(1024 * 1024),
-            kdf_iterations: Some(1),
-            kdf_parallelism: Some(1),
-            description: None,
-        };
-
-        let service = KdbxService::new();
-        service
-            .create_database(
-                &db_path_str,
-                Some("testpass"),
-                None,
-                "Favicon Tests",
-                &options,
-            )
-            .expect("create db");
-        let info = service.get_info(&db_path_str).expect("database info");
-
-        let entry_a = service
-            .create_entry(
-                &db_path_str,
-                &info.root_group_id,
-                CreateEntryData {
-                    title: "Entry A".to_string(),
-                    username: "alice".to_string(),
-                    password: SecureString::from("secret"),
-                    url: None,
-                    notes: None,
-                    icon_id: Some(0),
-                    tags: None,
-                    custom_fields: None,
-                    protected_custom_fields: None,
-                },
-            )
-            .expect("create entry A");
-        let entry_b = service
-            .create_entry(
-                &db_path_str,
-                &info.root_group_id,
-                CreateEntryData {
-                    title: "Entry B".to_string(),
-                    username: "bob".to_string(),
-                    password: SecureString::from("secret"),
-                    url: None,
-                    notes: None,
-                    icon_id: Some(0),
-                    tags: None,
-                    custom_fields: None,
-                    protected_custom_fields: None,
-                },
-            )
-            .expect("create entry B");
-
-        (service, dir, db_path_str, entry_a.id, entry_b.id)
-    }
-
-    fn tiny_png() -> Vec<u8> {
-        let image = ImageBuffer::from_pixel(2, 2, Rgba([0_u8, 128, 255, 255]));
-        let mut output = Cursor::new(Vec::new());
-        image
-            .write_to(&mut output, ImageFormat::Png)
-            .expect("write png");
-        output.into_inner()
-    }
-
-    fn serve_once(response: Vec<u8>) -> (String, thread::JoinHandle<()>) {
-        let listener = TcpListener::bind("127.0.0.1:0").expect("bind test server");
-        let address = listener.local_addr().expect("read local addr");
-        let handle = thread::spawn(move || {
-            let (mut stream, _) = listener.accept().expect("accept request");
-            let mut request = [0_u8; 1024];
-            let _ = stream.read(&mut request);
-            // Client may close the connection mid-write (e.g. when the
-            // streaming cap rejects an oversized body), so a broken pipe
-            // here is expected.
-            let _ = stream.write_all(&response);
-        });
-
-        (format!("http://{address}/favicon.ico"), handle)
-    }
-
-    #[test]
-    fn build_candidates_tries_hostname_before_root_domain() {
-        let hosts = FaviconLookup {
-            exact_authority: "app.example.com".to_string(),
-            hostname: "app.example.com".to_string(),
-            root_domain: Some("example.com".to_string()),
-        };
-        let candidates = build_favicon_candidates(&hosts, false);
-        let urls: Vec<&str> = candidates
-            .iter()
-            .map(|candidate| candidate.fetch_url.as_str())
-            .collect();
-
-        assert_eq!(urls.len(), 2);
-        assert_eq!(urls[0], "https://app.example.com/favicon.ico");
-        assert_eq!(urls[1], "https://example.com/favicon.ico");
-    }
-
-    #[test]
-    fn build_candidates_adds_opt_in_third_party_sources_for_public_domains() {
-        let hosts = FaviconLookup {
-            exact_authority: "app.example.com".to_string(),
-            hostname: "app.example.com".to_string(),
-            root_domain: Some("example.com".to_string()),
-        };
-        let candidates = build_favicon_candidates(&hosts, true);
-
-        assert!(candidates.iter().any(|candidate| {
-            candidate.fetch_url == "https://www.google.com/s2/favicons?domain=example.com&sz=64"
-        }));
-        assert!(candidates
-            .iter()
-            .any(|candidate| candidate.fetch_url == "https://icon.horse/icon/example.com"));
-        assert!(!candidates.iter().any(|candidate| {
-            candidate.fetch_url.contains("domain=app.example.com")
-                || candidate.fetch_url.ends_with("/app.example.com")
-        }));
-    }
-
-    #[test]
-    fn build_candidates_omits_third_party_sources_for_private_hosts() {
-        let hosts = FaviconLookup {
-            exact_authority: "nas.local:5001".to_string(),
-            hostname: "nas.local".to_string(),
-            root_domain: None,
-        };
-        let candidates = build_favicon_candidates(&hosts, true);
-        let urls: Vec<&str> = candidates
-            .iter()
-            .map(|candidate| candidate.fetch_url.as_str())
-            .collect();
-
-        assert_eq!(urls, vec!["https://nas.local:5001/favicon.ico"]);
-    }
-
-    #[test]
-    fn fetch_entry_favicon_returns_not_found_without_url() {
-        let (service, _dir, db_path, entry_id, _entry_b) = create_test_database();
-        let outcome = tauri::async_runtime::block_on(
-            service.fetch_entry_favicon(&db_path, &entry_id, false, true),
-        )
-        .expect("fetch favicon");
-
-        assert_eq!(outcome, FaviconFetchOutcome::NotFound);
-    }
-
-    #[test]
-    fn auto_fetch_does_not_refresh_cooldown_for_skipped_domain() {
-        let (service, _dir, db_path, entry_id, _entry_b) = create_test_database();
-        service
-            .update_entry(
-                &db_path,
-                &entry_id,
-                crate::dto::entry::UpdateEntryData {
-                    title: None,
-                    username: None,
-                    password: None,
-                    url: Some("https://example.com".to_string()),
-                    notes: None,
-                    icon_id: None,
-                    tags: None,
-                    custom_fields: None,
-                    protected_custom_fields: None,
-                },
-            )
-            .expect("set url");
-
-        let initial_stamp = std::time::Instant::now()
-            .checked_sub(std::time::Duration::from_secs(30))
-            .expect("build initial cooldown stamp");
-        {
-            let mut failures = service
-                .favicon_failed_domains
-                .lock()
-                .expect("lock cooldown map");
-            failures.insert("example.com".to_string(), initial_stamp);
-        }
-
-        // Non-forced (auto-fetch) path must honor the cooldown and must not
-        // bump the timestamp of a domain it never contacted.
-        let outcome = tauri::async_runtime::block_on(
-            service.fetch_entry_favicon(&db_path, &entry_id, false, false),
-        )
-        .expect("fetch favicon");
-
-        assert_eq!(outcome, FaviconFetchOutcome::NotFound);
-
-        let failures = service
-            .favicon_failed_domains
-            .lock()
-            .expect("lock cooldown map");
-        let stamp = failures
-            .get("example.com")
-            .copied()
-            .expect("cooldown entry preserved");
-        assert_eq!(
-            stamp, initial_stamp,
-            "auto-fetch must not refresh a cooldown stamp for a skipped domain"
-        );
-    }
-
-    #[test]
-    fn manual_force_fetch_bypasses_cooldown_and_attempts_the_domain() {
-        let (service, _dir, db_path, entry_id, _entry_b) = create_test_database();
-        service
-            .update_entry(
-                &db_path,
-                &entry_id,
-                crate::dto::entry::UpdateEntryData {
-                    title: None,
-                    username: None,
-                    password: None,
-                    url: Some("https://example.com".to_string()),
-                    notes: None,
-                    icon_id: None,
-                    tags: None,
-                    custom_fields: None,
-                    protected_custom_fields: None,
-                },
-            )
-            .expect("set url");
-
-        let initial_stamp = std::time::Instant::now()
-            .checked_sub(std::time::Duration::from_secs(30))
-            .expect("build initial cooldown stamp");
-        {
-            let mut failures = service
-                .favicon_failed_domains
-                .lock()
-                .expect("lock cooldown map");
-            failures.insert("example.com".to_string(), initial_stamp);
-        }
-
-        // Force=true mirrors the manual "Refetch favicon" path. The cooldown
-        // gate must be bypassed so the user actually retries the host. The
-        // attempt then fails (no network in tests) and refreshes the stamp.
-        let outcome = tauri::async_runtime::block_on(
-            service.fetch_entry_favicon(&db_path, &entry_id, false, true),
-        )
-        .expect("fetch favicon");
-
-        assert_eq!(outcome, FaviconFetchOutcome::NotFound);
-
-        let failures = service
-            .favicon_failed_domains
-            .lock()
-            .expect("lock cooldown map");
-        let stamp = failures
-            .get("example.com")
-            .copied()
-            .expect("cooldown entry preserved");
-        assert!(
-            stamp > initial_stamp,
-            "force=true must contact the domain, which refreshes the failure stamp on miss"
-        );
-    }
-
-    #[test]
-    fn fetch_entry_favicon_returns_not_found_for_invalid_url() {
-        let (service, _dir, db_path, entry_id, _entry_b) = create_test_database();
-        service
-            .update_entry(
-                &db_path,
-                &entry_id,
-                crate::dto::entry::UpdateEntryData {
-                    title: None,
-                    username: None,
-                    password: None,
-                    url: Some("not-a-valid-url".to_string()),
-                    notes: None,
-                    icon_id: None,
-                    tags: None,
-                    custom_fields: None,
-                    protected_custom_fields: None,
-                },
-            )
-            .expect("set invalid url");
-
-        let outcome = tauri::async_runtime::block_on(
-            service.fetch_entry_favicon(&db_path, &entry_id, false, true),
-        )
-        .expect("fetch favicon");
-
-        assert_eq!(outcome, FaviconFetchOutcome::NotFound);
-    }
-
-    #[test]
-    fn set_entry_custom_icon_assigns_existing_icon() {
-        let (service, _dir, db_path, entry_a, entry_b) = create_test_database();
-        let icon_bytes = [0x89, b'P', b'N', b'G', 99];
-        service
-            .assign_entry_custom_icon(&db_path, &entry_a, &icon_bytes, "image/png", true)
-            .expect("seed icon");
-
-        let icon_uuid = {
-            let normalized = KdbxService::normalize_path(&db_path);
-            let databases = service.lock_databases().expect("lock databases");
-            let open_db = databases.get(&normalized).expect("open db");
-            let db = open_db.db_or_locked().expect("unlocked db");
-            let icon_uuid = db
-                .iter_all_custom_icons()
-                .next()
-                .expect("custom icon exists")
-                .id()
-                .uuid()
-                .to_string();
-            icon_uuid
-        };
-
-        let changed = service
-            .set_entry_custom_icon(&db_path, &entry_b, &icon_uuid)
-            .expect("set icon on b");
-        assert!(changed);
-
-        let unchanged = service
-            .set_entry_custom_icon(&db_path, &entry_b, &icon_uuid)
-            .expect("set icon on b again");
-        assert!(!unchanged);
-    }
-
-    #[test]
-    fn set_entry_custom_icon_rejects_unknown_uuid() {
-        let (service, _dir, db_path, entry_a, _entry_b) = create_test_database();
-        let result = service.set_entry_custom_icon(
-            &db_path,
-            &entry_a,
-            "00000000-0000-0000-0000-000000000000",
-        );
-
-        assert!(matches!(result, Err(AppError::InvalidInput(_))));
-    }
-
-    #[test]
-    fn extract_hosts_normalizes_subdomain_and_root_domain() {
-        let hosts = extract_hosts("https://APP.Example.COM/login").expect("extract hosts");
-
-        assert_eq!(hosts.exact_authority, "app.example.com");
-        assert_eq!(hosts.hostname, "app.example.com");
-        assert_eq!(hosts.root_domain.as_deref(), Some("example.com"));
-        assert_eq!(
-            extract_hosts("https://login.example.co.uk")
-                .expect("extract multi-label public suffix root")
-                .root_domain
-                .as_deref(),
-            Some("example.co.uk")
-        );
-
-        let bare = extract_hosts("https://example.co.uk").expect("registrable apex");
-        assert_eq!(bare.hostname, "example.co.uk");
-        assert!(
-            bare.root_domain.is_none(),
-            "the registrable apex has no further root to fall back to"
-        );
-
-        let ip = extract_hosts("http://127.0.0.1:3000").expect("ip literal");
-        assert_eq!(ip.exact_authority, "127.0.0.1:3000");
-        assert_eq!(ip.hostname, "127.0.0.1");
-        assert!(
-            ip.root_domain.is_none(),
-            "ip literals never produce a root host"
-        );
-
-        assert_eq!(extract_hosts("not a url"), None);
-    }
-
-    #[test]
-    fn extract_hosts_preserves_explicit_ports_for_exact_fetches() {
-        let hosts = extract_hosts("https://app.example.com:8443/login").expect("extract hosts");
-
-        assert_eq!(hosts.exact_authority, "app.example.com:8443");
-        assert_eq!(hosts.hostname, "app.example.com");
-        assert_eq!(hosts.root_domain.as_deref(), Some("example.com"));
-
-        let candidates = build_favicon_candidates(&hosts, true);
-        let urls: Vec<&str> = candidates
-            .iter()
-            .map(|candidate| candidate.fetch_url.as_str())
-            .collect();
-
-        assert_eq!(urls[0], "https://app.example.com:8443/favicon.ico");
-        assert_eq!(urls[1], "https://example.com/favicon.ico");
-        assert!(urls.contains(&"https://www.google.com/s2/favicons?domain=example.com&sz=64"));
-        assert!(!urls
-            .iter()
-            .skip(2)
-            .any(|url| url.contains("app.example.com:8443")));
-    }
-
-    #[test]
-    fn extract_hosts_rejects_non_http_schemes() {
-        // Non-web entries (ssh, db, ldap, etc.) must not trigger favicon
-        // network activity. extract_hosts returns None so fetch_entry_favicon
-        // short-circuits with NotFound before any DNS or HTTP traffic.
-        for url in [
-            "ssh://bastion.corp.example/",
-            "postgres://db.internal:5432/prod",
-            "ldap://corp-ad.internal/",
-            "file:///etc/hosts",
-            "ftp://files.example.com/",
-            "obsidian://open?vault=Notes",
-        ] {
-            assert_eq!(extract_hosts(url), None, "expected no fetch for {url}");
-        }
-    }
-
-    #[test]
-    fn extract_hosts_skips_root_fallback_for_psl_hosting_platforms() {
-        // PSL entries like github.io / netlify.app / vercel.app are public
-        // suffixes themselves, so user.github.io is already the registrable
-        // apex. The root-host fallback must be skipped so we never fetch
-        // the provider's marketing-page favicon by accident.
-        for host in [
-            "user.github.io",
-            "team.netlify.app",
-            "demo.vercel.app",
-            "tenant.azurewebsites.net",
-        ] {
-            let url = format!("https://{host}/");
-            let parsed = extract_hosts(&url).expect("parse psl host");
-            assert_eq!(parsed.hostname, host);
-            assert!(
-                parsed.root_domain.is_none(),
-                "{host} should not produce a root-host fallback"
-            );
-        }
-    }
-
-    #[test]
-    fn detect_icon_mime_recognizes_supported_signatures() {
-        assert_eq!(
-            detect_icon_mime(b"  \n<svg viewBox=\"0 0 1 1\" />"),
-            "image/svg+xml"
-        );
-        assert_eq!(detect_icon_mime(&[0x89, b'P', b'N', b'G']), "image/png");
-        assert_eq!(detect_icon_mime(&[0xFF, 0xD8, 0xFF]), "image/jpeg");
-        assert_eq!(detect_icon_mime(b"GIF89a"), "image/gif");
-        assert_eq!(detect_icon_mime(&[0x00, 0x00, 0x01, 0x00]), "image/x-icon");
-        assert_eq!(detect_icon_mime(b"BM..."), "image/bmp");
-        assert_eq!(detect_icon_mime(b"RIFF....WEBP"), "image/webp");
-        assert_eq!(detect_icon_mime(&[0x49, 0x49, 0x2A, 0x00]), "image/tiff");
-        assert_eq!(detect_icon_mime(b"plain text"), "application/octet-stream");
-    }
-
-    #[test]
-    fn normalize_favicon_bytes_resizes_decodable_images_to_png() {
-        let original = tiny_png();
-        let (normalized, mime_type) =
-            normalize_favicon_bytes(&original, Some("image/png".to_string()));
-
-        assert_eq!(mime_type, "image/png");
-        assert!(normalized.starts_with(&[0x89, b'P', b'N', b'G']));
-        assert!(
-            image::load_from_memory(&normalized).is_ok(),
-            "normalized favicon should remain a decodable image"
-        );
-    }
-
-    #[test]
-    fn normalize_favicon_bytes_preserves_svg_with_content_type() {
-        let svg = br#"<svg xmlns="http://www.w3.org/2000/svg"></svg>"#;
-        let (normalized, mime_type) =
-            normalize_favicon_bytes(svg, Some("image/svg+xml".to_string()));
-
-        assert_eq!(normalized, svg);
-        assert_eq!(mime_type, "image/svg+xml");
-    }
-
-    #[test]
-    fn fetch_favicon_bytes_accepts_valid_image_responses() {
-        let png = tiny_png();
-        let response = [
-            b"HTTP/1.1 200 OK\r\nContent-Type: image/png; charset=utf-8\r\nContent-Length: "
-                .as_slice(),
-            png.len().to_string().as_bytes(),
-            b"\r\n\r\n",
-            png.as_slice(),
-        ]
-        .concat();
-        let (url, handle) = serve_once(response);
-        let client = reqwest::Client::new();
-
-        let fetched =
-            tauri::async_runtime::block_on(fetch_favicon_bytes(&client, &url)).expect("fetch png");
-        handle.join().expect("server finishes");
-
-        assert_eq!(fetched.0, png);
-        assert_eq!(fetched.1.as_deref(), Some("image/png"));
-    }
-
-    #[test]
-    fn fetch_favicon_bytes_rejects_non_image_content_type() {
-        let body = b"<html>not an icon</html>";
-        let response = [
-            b"HTTP/1.1 200 OK\r\nContent-Type: text/html\r\nContent-Length: ".as_slice(),
-            body.len().to_string().as_bytes(),
-            b"\r\n\r\n",
-            body.as_slice(),
-        ]
-        .concat();
-        let (url, handle) = serve_once(response);
-        let client = reqwest::Client::new();
-
-        let fetched = tauri::async_runtime::block_on(fetch_favicon_bytes(&client, &url));
-        handle.join().expect("server finishes");
-
-        assert!(fetched.is_none());
-    }
-
-    #[test]
-    fn fetch_favicon_bytes_caps_streamed_body_without_content_length() {
-        // No Content-Length header — the streaming reader must enforce the
-        // cap on its own and short-circuit before buffering the whole body.
-        let oversized = vec![0_u8; FAVICON_MAX_BYTES + 1];
-        let response = [
-            b"HTTP/1.1 200 OK\r\nContent-Type: image/png\r\nTransfer-Encoding: chunked\r\n\r\n"
-                .as_slice(),
-            format!("{:x}\r\n", oversized.len()).as_bytes(),
-            oversized.as_slice(),
-            b"\r\n0\r\n\r\n",
-        ]
-        .concat();
-        let (url, handle) = serve_once(response);
-        let client = reqwest::Client::new();
-
-        let fetched = tauri::async_runtime::block_on(fetch_favicon_bytes(&client, &url));
-        let _ = handle.join();
-
-        assert!(fetched.is_none());
-    }
-
-    #[test]
-    fn fetch_favicon_bytes_rejects_non_https_redirects() {
-        // Server replies with a redirect to an http:// target. The redirect
-        // policy should stop following, so the final response stays 302 and
-        // is_success() filters it out.
-        let response = b"HTTP/1.1 302 Found\r\nLocation: http://example.invalid/favicon.ico\r\nContent-Length: 0\r\n\r\n".to_vec();
-        let (url, handle) = serve_once(response);
-        let client = reqwest::Client::builder()
-            .redirect(Policy::custom(|attempt| {
-                if attempt.url().scheme() != "https" {
-                    attempt.stop()
-                } else if attempt.previous().len() >= 5 {
-                    attempt.error("too many redirects")
-                } else {
-                    attempt.follow()
-                }
-            }))
-            .build()
-            .expect("client");
-
-        let fetched = tauri::async_runtime::block_on(fetch_favicon_bytes(&client, &url));
-        handle.join().expect("server finishes");
-
-        assert!(fetched.is_none());
-    }
-
-    #[test]
-    fn fetch_favicon_bytes_rejects_oversized_icons() {
-        let oversized = vec![0_u8; FAVICON_MAX_BYTES + 1];
-        let response = [
-            b"HTTP/1.1 200 OK\r\nContent-Type: image/png\r\nContent-Length: ".as_slice(),
-            oversized.len().to_string().as_bytes(),
-            b"\r\n\r\n",
-            oversized.as_slice(),
-        ]
-        .concat();
-        let (url, handle) = serve_once(response);
-        let client = reqwest::Client::new();
-
-        let fetched = tauri::async_runtime::block_on(fetch_favicon_bytes(&client, &url));
-        handle.join().expect("server finishes");
-
-        assert!(fetched.is_none());
-    }
-
-    #[test]
-    fn entry_with_custom_icon_reports_no_builtin_icon_id() {
-        // keepass 0.12 made builtin and custom icons mutually exclusive on
-        // Entry. convert_entry must reflect that honestly: when a custom
-        // icon is set, iconId is null and customIconUuid carries the UUID.
-        // The previous "always emit iconId=0" workaround caused the
-        // frontend to echo iconId=0 back on the next save and silently
-        // destroy the custom icon — see
-        // update_entry_without_icon_id_preserves_custom_icon.
-        let (service, _dir, db_path, entry_a, _entry_b) = create_test_database();
-        let icon_bytes = [0x89, b'P', b'N', b'G', 0x0D, 0x0A, 0x1A, 0x0A, 7];
-
-        service
-            .assign_entry_custom_icon(&db_path, &entry_a, &icon_bytes, "image/png", true)
-            .expect("assign favicon");
-
-        let entries = service.list_entries(&db_path, None).expect("list entries");
-        let entry = entries
-            .iter()
-            .find(|e| e.id == entry_a)
-            .expect("entry returned by list_entries");
-        assert!(
-            entry.icon_id.is_none(),
-            "iconId must be None when the entry has only a custom icon"
-        );
-        assert!(
-            entry.custom_icon_uuid.is_some(),
-            "customIconUuid must round-trip"
-        );
-    }
-
-    #[test]
-    fn update_entry_without_icon_id_preserves_custom_icon() {
-        // Regression: editing a non-icon field on a favicon-bearing entry
-        // used to destroy the favicon because the frontend echoed back
-        // iconId=0 and set_icon_builtin(0) cleared Icon::Custom. Sending
-        // icon_id: None must leave the entry's icon untouched.
-        let (service, _dir, db_path, entry_a, _entry_b) = create_test_database();
-        let icon_bytes = [0x89, b'P', b'N', b'G', 0x0D, 0x0A, 0x1A, 0x0A, 9];
-        service
-            .assign_entry_custom_icon(&db_path, &entry_a, &icon_bytes, "image/png", true)
-            .expect("assign favicon");
-
-        service
-            .update_entry(
-                &db_path,
-                &entry_a,
-                crate::dto::entry::UpdateEntryData {
-                    title: Some("Renamed".to_string()),
-                    username: None,
-                    password: None,
-                    url: None,
-                    notes: None,
-                    icon_id: None,
-                    tags: None,
-                    custom_fields: None,
-                    protected_custom_fields: None,
-                },
-            )
-            .expect("update title");
-
-        let entry = service.get_entry(&db_path, &entry_a).expect("get entry");
-        assert!(
-            entry.custom_icon_uuid.is_some(),
-            "favicon must survive an unrelated field update"
-        );
-        assert!(
-            entry.icon_id.is_none(),
-            "the entry has no builtin icon and the update didn't touch one"
-        );
-    }
-
-    #[test]
-    fn update_entry_with_builtin_zero_switches_off_custom_icon() {
-        // Picker-to-Key (icon 0) path on a custom-icon entry: the frontend
-        // sends iconId=Some(0) (dirty) and the customIconUuid clear is
-        // applied separately. The backend must end in Icon::BuiltIn(0).
-        let (service, _dir, db_path, entry_a, _entry_b) = create_test_database();
-        let icon_bytes = [0x89, b'P', b'N', b'G', 0x0D, 0x0A, 0x1A, 0x0A, 10];
-        service
-            .assign_entry_custom_icon(&db_path, &entry_a, &icon_bytes, "image/png", true)
-            .expect("assign favicon");
-
-        service
-            .update_entry(
-                &db_path,
-                &entry_a,
-                crate::dto::entry::UpdateEntryData {
-                    title: None,
-                    username: None,
-                    password: None,
-                    url: None,
-                    notes: None,
-                    icon_id: Some(0),
-                    tags: None,
-                    custom_fields: None,
-                    protected_custom_fields: None,
-                },
-            )
-            .expect("update icon to builtin 0");
-
-        let entry = service.get_entry(&db_path, &entry_a).expect("get entry");
-        assert_eq!(entry.icon_id, Some(0));
-        assert!(
-            entry.custom_icon_uuid.is_none(),
-            "switching to builtin 0 must drop the custom icon"
-        );
-    }
-
-    #[test]
-    fn assign_entry_custom_icon_deduplicates_icon_bytes() {
-        let (service, _dir, db_path, entry_a, entry_b) = create_test_database();
-        let icon_bytes = [0x89, b'P', b'N', b'G', 0x0D, 0x0A, 0x1A, 0x0A, 0x00];
-
-        let changed_a = service
-            .assign_entry_custom_icon(&db_path, &entry_a, &icon_bytes, "image/png", true)
-            .expect("assign icon A");
-        let changed_b = service
-            .assign_entry_custom_icon(&db_path, &entry_b, &icon_bytes, "image/png", true)
-            .expect("assign icon B");
-
-        assert!(changed_a);
-        assert!(changed_b);
-
-        let normalized = KdbxService::normalize_path(&db_path);
-        let databases = service.lock_databases().expect("lock databases");
-        let open_db = databases.get(&normalized).expect("open db");
-        let db = open_db.db_or_locked().expect("unlocked db");
-        assert_eq!(db.iter_all_custom_icons().count(), 1);
-
-        let icon_for = |entry_id: &str| -> uuid::Uuid {
-            let eid = find_entry_id(db, entry_id).expect("entry id");
-            let entry = db.entry(eid).expect("entry ref");
-            match entry.icon() {
-                Some(Icon::Custom(cid)) => cid.uuid(),
-                other => unreachable!("entry {entry_id} has no custom icon: {other:?}"),
-            }
-        };
-        assert_eq!(icon_for(&entry_a), icon_for(&entry_b));
-    }
-
-    #[test]
-    fn clear_entry_custom_icon_detaches_icon_uuid() {
-        let (service, _dir, db_path, entry_a, _entry_b) = create_test_database();
-        let icon_bytes = [1_u8, 2, 3, 4, 5];
-        service
-            .assign_entry_custom_icon(
-                &db_path,
-                &entry_a,
-                &icon_bytes,
-                "application/octet-stream",
-                true,
-            )
-            .expect("assign icon");
-
-        let cleared = service
-            .clear_entry_custom_icon(&db_path, &entry_a)
-            .expect("clear icon");
-        let cleared_again = service
-            .clear_entry_custom_icon(&db_path, &entry_a)
-            .expect("clear icon again");
-
-        assert!(cleared);
-        assert!(!cleared_again);
-
-        let normalized = KdbxService::normalize_path(&db_path);
-        let databases = service.lock_databases().expect("lock databases");
-        let open_db = databases.get(&normalized).expect("open db");
-        let db = open_db.db_or_locked().expect("unlocked db");
-        let eid = find_entry_id(db, &entry_a).expect("entry id");
-        let entry = db.entry(eid).expect("entry ref");
-        assert!(!matches!(entry.icon(), Some(Icon::Custom(_))));
-    }
-
-    #[test]
-    fn assign_entry_custom_icon_respects_existing_icon_without_force() {
-        let (service, _dir, db_path, entry_a, _entry_b) = create_test_database();
-        let first_icon = [0x89, b'P', b'N', b'G', 1];
-        let replacement_icon = [0x89, b'P', b'N', b'G', 2];
-
-        assert!(service
-            .assign_entry_custom_icon(&db_path, &entry_a, &first_icon, "image/png", true)
-            .expect("assign first icon"));
-        assert!(
-            !service
-                .assign_entry_custom_icon(&db_path, &entry_a, &replacement_icon, "image/png", false)
-                .expect("skip replacement"),
-            "non-forced favicon fetches should preserve a user-selected icon"
-        );
-
-        let normalized = KdbxService::normalize_path(&db_path);
-        let databases = service.lock_databases().expect("lock databases");
-        let open_db = databases.get(&normalized).expect("open db");
-        let db = open_db.db_or_locked().expect("unlocked db");
-        let mut icons: Vec<Vec<u8>> = db
-            .iter_all_custom_icons()
-            .map(|icon| icon.data.clone())
-            .collect();
-        assert_eq!(icons.len(), 1);
-        assert_eq!(icons.pop().expect("icon data"), first_icon);
-    }
-}
